Align remaining review nits for delay validation and scheduler typing
Agent-Logs-Url: https://github.com/jiang223/ldfw/sessions/af09e203-0ecd-4401-8810-283338ade7c8

+++ core/task_scheduler.py
@@ -11,7 +11,7 @@ MIN_TASK_INTERVAL_SEC = 0.1
 @dataclass
 class ScheduledTask:
     name: str
-    interval_sec: int
+    interval_sec: float
     callback: Callable[[], None]
     enabled: bool = True
